pyproject.toml
[project]
name = "statstonk"
version = "0.1.0"
description = "Football betting analytics platform — Serie A and top European leagues"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "playwright>=1.40",
    "pandas>=2.1",
    "numpy>=1.26",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "apscheduler>=3.10",
    "python-telegram-bot>=20.7",
    "streamlit>=1.30",
    "python-dotenv>=1.0",
    "plotly>=5.18",
    "httpx>=0.26",
    "tenacity>=8.2",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.3",
    "mypy>=1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scrapers", "models", "analytics", "alerts", "scheduler", "dashboard"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
